move modules to adhere with src layout
