+++ rust-installer/src/openvair_manager/cmd_runner.rs
