+++ rust-installer/src/openvair_manager/docker.rs
