+++ rust-installer/src/openvair_manager/pkg_management.rs
@@ -1,4 +1,4 @@
-use crate::cmd_runner::{CommandResult, CommandRunner};
+use crate::openvair_manager::cmd_runner::{CommandResult, CommandRunner};
 use std::{process::Command, rc::Rc};
 
 pub trait PackageProvider {
