+++ rust-installer/src/openvair_manager/pkg_management/tests.rs
