+++ rust-installer/src/openvair_manager/prometheus.rs
@@ -1,0 +1,1 @@
+pub mod installer;
